refactor(e2e): reorganize test files by user flow
Move Test_ClientVersion to smoke_test.go, merge Test_DeleteAccount into
sign_up_test.go, drop redundant Test_PlayerManagement, add flow doc comments.
